fix: load-more button added in place of infinite scroll
fix: accessibility issue resolved assigning aria-levels to UI components

fix: learner view loading changes and ui fix

fix: removed unused component

fix: lint fixes
