+++ src/discussions/learners/messages.js
@@ -13,10 +13,10 @@ const messages = defineMessages({
     id: 'discussions.learner.lastLogin',
     defaultMessage: 'Last active {lastActiveTime}',
   },
-  loadMorePosts: {
-    id: 'discussions.learner.loadMostPosts',
-    defaultMessage: 'Load more posts',
-    description: 'Text on button for loading more posts by a user',
+  loadMore: {
+    id: 'discussions.learner.loadMostLearners',
+    defaultMessage: 'Load more',
+    description: 'Text on button for loading more learners',
   },
   back: {
     id: 'discussions.learner.back',
