+++ src/discussions/messages.js
@@ -143,6 +143,11 @@ const messages = defineMessages({
     defaultMessage: 'TA',
     description: 'A label for community TAs displayed next to their username.',
   },
+  loadMorePosts: {
+    id: 'discussions.learner.loadMostPosts',
+    defaultMessage: 'Load more posts',
+    description: 'Text on button for loading more posts by a user',
+  },
 });
 
 export default messages;
